+++ CHANGELOG.md
@@ -35,6 +35,19 @@ And a few potential drawbacks to lookout for:
  **BugFixes**:
  - Source info not properly read from external storage NAS #1727
 
+## v1.1.5-beta
+
+ **Notes**:
+ - Major git container tag request #1756
+
+ **BugFixes**:
+ - Sharing broken if source `disableIndexing: true` (#1742)
+ - Password protected share permission issues (#1729) (#1606) (#1593)
+ - Sidebar Sliding left (#1737)
+ - file permissions not working (#1762)
+ - handle Close() errors in archive creation (#1745)
+ - Move custom CSS and user-selected theme styles to the end of the body (#1744)
+
 ## v1.1.4-beta
 
  **New Features**: